fix references from previous DataGrid reorganization
