+++ src/routes/dictionaries/index.svelte
@@ -6,7 +6,7 @@
   import Chip from '../../components/ui/Chip.svelte';
   import CssGrid from '../../components/ui/CssGrid.svelte';
   import DataGrid from '../../components/ui/DataGrid/DataGrid.svelte';
-  import DataGridActions from '../../components/ui/DataGridActions.svelte';
+  import DataGridActions from '../../components/ui/DataGrid/DataGridActions.svelte';
   import Panel from '../../components/ui/Panel.svelte';
   import { createDictionaryError, creatingDictionary, sortedDictionaries } from '../../stores/expansion';
   import effects from '../../utilities/effects';
