+++ src/routes/plans/index.svelte
@@ -12,7 +12,7 @@
   import Chip from '../../components/ui/Chip.svelte';
   import CssGrid from '../../components/ui/CssGrid.svelte';
   import DataGrid from '../../components/ui/DataGrid/DataGrid.svelte';
-  import DataGridActions from '../../components/ui/DataGridActions.svelte';
+  import DataGridActions from '../../components/ui/DataGrid/DataGridActions.svelte';
   import Panel from '../../components/ui/Panel.svelte';
   import { field } from '../../stores/form';
   import { createPlanError, creatingPlan } from '../../stores/plan';
